ALGOR-14: migrate to the new Tinkoff Invest API

+++ src/main/scala/com/github/ppotseluev/algorate/model/package.scala
@@ -6,6 +6,6 @@ package object model {
   type InstrumentId = String @@ Tags.InstrumentId
   type BrokerAccountId = String @@ Tags.BrokerAccountId
   type OrderId = String @@ Tags.OrderId
-  type Price = Double @@ Tags.Price
+  type Price = BigDecimal @@ Tags.Price
   type Ticker = String @@ Tags.Ticker
 }
